pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.2",
    "python-binance>=1.0.29",
    "python-dotenv>=1.1.1",
]
